third_party/blink/web_tests/external/wpt/magnetometer/idlharness.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] UncalibratedMagnetometer interface: existence and properties of interface object
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface object length
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface object name
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: attribute x
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: attribute y
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: attribute z
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: attribute xBias
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: attribute yBias
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer interface: attribute zBias
  assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
[FAIL] UncalibratedMagnetometer must be primary interface of new UncalibratedMagnetometer();
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
[FAIL] Stringification of new UncalibratedMagnetometer();
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
[FAIL] UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "x" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
[FAIL] UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "y" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
[FAIL] UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "z" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
[FAIL] UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "xBias" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
[FAIL] UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "yBias" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
[FAIL] UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "zBias" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
Harness: the test ran to completion.
